240130 [은지/오류수정] 오픈뱅킹API / 계좌내역이 존재하는 계좌 삭제 시 오류 수정
삭제를 진행하게 되면, 계좌내역테이블이 계좌테이블의 account를 참조하고 있기 때문에 삭제가 진행되지 않아 오류가 뜨게 됨. 이 때, "삭제할 수 없습니다" 라는 alert를 띄우고 다시 /develop/openbank 로 돌아가도록 수정해줌.

+++ src/main/java/com/cruise/project_cruise/controller/openbank/OpenBankController.java
@@ -1,10 +1,7 @@
 package com.cruise.project_cruise.controller.openbank;
 
 import com.cruise.project_cruise.dto.develop.OpenBankDTO;
-import com.cruise.project_cruise.service.CrewBoardService;
-import com.cruise.project_cruise.service.CrewCommentService;
-import com.cruise.project_cruise.service.DevelopOpenBankingService;
-import com.cruise.project_cruise.service.TemplateService;
+import com.cruise.project_cruise.service.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
@@ -19,15 +16,19 @@ public class OpenBankController {
 
     @Autowired
     private DevelopOpenBankingService developOpenBankingService;
+    @Autowired
+    private DevelopOpenBankUsingService developOpenBankUsingService;
 
     @RequestMapping(value="")
     public ModelAndView bankAccountList() throws Exception {
 
         ModelAndView mav = new ModelAndView();
+        String alertMsg = "";
 
         List<OpenBankDTO> accountList =
                 developOpenBankingService.getAccountList();
 
+        mav.addObject("alertMsg", alertMsg);
         mav.addObject("accountList",accountList);
         mav.setViewName("forDevelop/account");
 
@@ -56,9 +57,19 @@ public class OpenBankController {
     @PostMapping(value="/deleteAccount")
     public ModelAndView bankAccountDelete(@RequestParam String account) throws Exception {
         ModelAndView mav = new ModelAndView();
+        String alertMsg = "";
+
+        if(!developOpenBankUsingService.getUsingList(account).isEmpty()) {
+            alertMsg = "계좌내역이 존재하는 계좌는 삭제할 수 없습니다.";
+            mav.addObject("alertMsg",alertMsg);
+            mav.setViewName("forDevelop/account");
+            return mav;
+        }
 
         developOpenBankingService.deleteAccount(account);
         System.out.println("[Develop] 오픈뱅킹 테스트 계좌 Delete : " + account);
+
+        mav.addObject(alertMsg);
         mav.setViewName("redirect: ");
         return mav;
     }
